fixed: update model
