+++ src/shared/types/network-types.ts
@@ -174,6 +174,16 @@ export interface BlockchainResponse {
   color: string;
 }
 
+export interface WalletType {
+  name: string;
+  icon: string;
+  address: string;
+  chain_id: string;
+  id: number;
+  coins: string[];
+  color: string;
+}
+
 interface Thumbnail {
   url: string;
 }
