+++ src/shared/types/wallet-types.ts
@@ -3,3 +3,14 @@ export type FlowAddress = `0x${string & { length: 16 }}` | `${string & { length:
 
 // ActiveChildType is the type of the active child in the wallet. It can be 'evm', a FlowAddress, or null.
 export type ActiveChildType = 'evm' | FlowAddress | null;
+
+export interface CoinItem {
+  coin: string;
+  unit: string;
+  balance: number;
+  price: number;
+  change24h: number | null;
+  total: number;
+  icon: string;
+  custom?: boolean;
+}
